- now showing number of new photos after refresh - now showing no new photos were found if that's the case after refresh - again fixed thread issue with queries

+++ app/src/main/java/nl/acidcats/tumblrlikes/ui/screens/load_likes_screen/LoadLikesFragment.kt
@@ -61,7 +61,7 @@ class LoadLikesFragment : BaseFragment(), LoadLikesScreenContract.View {
         }
     }
 
-    override fun showAllLikesLoaded(count: Long) {
+    override fun showAllLikesLoaded(count: Int) {
         imageCountText.text = getString(string.total_image_count, count)
         loadingText.text = getString(string.all_loaded)
     }
