+++ app/src/main/java/nl/acidcats/tumblrlikes/ui/screens/load_likes_screen/LoadLikesScreenContract.kt
@@ -25,7 +25,7 @@ interface LoadLikesScreenContract {
 
         fun showErrorAlert(errorStringId: Int)
 
-        fun showAllLikesLoaded(count: Long)
+        fun showAllLikesLoaded(count: Int)
 
         fun showLoadingCancelled()
     }
