+++ datalib/src/main/kotlin/nl/acidcats/tumblrlikes/core/usecases/likes/GetLikesUseCase.kt
@@ -8,5 +8,5 @@ import rx.subjects.BehaviorSubject
  * Created on 24/10/2018.
  */
 interface GetLikesUseCase {
-    fun loadAllLikes(mode: LoadLikesMode, loadingInterruptor: List<Boolean>, currentTimeInMs: Long, pageProgress: BehaviorSubject<Int>? = null): Observable<Long>
+    fun loadAllLikes(mode: LoadLikesMode, loadingInterruptor: List<Boolean>, currentTimeInMs: Long, pageProgress: BehaviorSubject<Int>? = null): Observable<Int>
 }
